@@ -26,25 +26,20 @@
   "author": "Gil Pedersen <gpdev@gpost.dk>",
   "license": "BSD-2-Clause",
   "dependencies": {
-    "@hapi/bounce": "^1.3.0",
-    "aws-sdk": "^2.179.0",
-    "commander": "^2.3.0",
+    "@hapi/bounce": "^3.0.0",
+    "commander": "^10.0.0",
     "debug": "^4.1.1",
-    "hls-segment-reader": "^5.0.0",
-    "m3u8parse": "^2.0.1",
+    "hls-playlist-reader": "^3.0.0-beta.1",
+    "hls-segment-reader": "file:../node-hls-segment-reader",
     "measured-core": "^1.41.0",
-    "mime-types": "^2.0.1",
-    "mkdirp": "^0.5.0",
-    "noptify": "0.0.3",
     "pati": "^1.1.2",
-    "readable-stream": "^3.3.0",
-    "stream-each": "^1.1.2",
     "udp-blast": "^2.0.0",
-    "uristream": "^3.0.0",
-    "write-file-atomic": "^2.3.0"
+    "uristream": "^6.3.1"
   },
-  "devDependencies": {},
   "engines": {
-    "node": ">=10.12.0"
+    "node": ">=14.17.0"
+  },
+  "devDependencies": {
+    "m3u8parse": "^4.0.0-rc4"
   }
 }
