@@ -1,13 +1,10 @@
-'use strict';
+import { PassThrough, Readable } from 'stream';
 
-const { PassThrough, Readable } = require('readable-stream');
+import Oncemore from 'oncemore';
+import Pati from 'pati';
 
-const StreamEach = require('stream-each');
-const Oncemore = require('oncemore');
-const Pati = require('pati');
-
-const TsSmooth = require('./tssmooth');
-const SegmentDecrypt = require('./segment-decrypt');
+import TsSmooth from './tssmooth.js';
+import SegmentDecrypt from './segment-decrypt.js';
 
 
 const internals = {
@@ -38,8 +35,12 @@ internals.pump = function(src, dst, done) {
 
 // TODO: use pipe as interface to segment-reader?
 
-const HlsReader = class extends Readable {
+export const HlsReader = class extends Readable {
 
+  /**
+   * @param {import("hls-segment-reader").HlsSegmentStreamer} segmentReader 
+   * @param {*} options 
+   */
   constructor(segmentReader, options) {
 
     options = options || {};
@@ -60,25 +61,34 @@ const HlsReader = class extends Readable {
 
     this.isReading = false;
     this.isHooked = false;
-    this.buffer = new PassThrough({ highWaterMark: this.bufferSize });
+    this.buffer = new PassThrough({ writableHighWaterMark: this.bufferSize, readableHighWaterMark: 0 });
+    this.presentDelay = 0;
 
-    const processStreams = () => {
+    this.reader.fetcher.source.index().then(async (/** @type {import("hls-playlist-reader/lib/fetcher").PlaylistObject} */ { index }) => {
 
-      StreamEach(this.reader, this.process.bind(this), (err) => {
+      this.presentDelay = index.target_duration * 3;
+      if (index.server_control) {
+        this.presentDelay = index.server_control.get('PART-HOLD-BACK', 'float') || index.server_control.get('HOLD-BACK', 'float') || this.presentDelay;
+      }
 
-        if (err) {
-          return processStreams(); // retry
+      try {
+        const reader = this.reader.getReader();
+        while (true) {
+          const { done, value } = await reader.read();
+          if (done) {
+            break;
+          }
+
+          await this.process(value);
         }
 
         this.buffer.end();
-      });
-    };
-
-    if (this.reader.index) {
-      processStreams();
-    } else {
-      this.reader.once('index', processStreams);
-    }
+      }
+      catch (err) {
+        console.error('FATAL error', err);   // TODO: retry when appropriate
+        //return processStreams(); // retry
+      }
+    });
 
     // start output if needed
     if (!this.sync) {
@@ -95,15 +105,17 @@ const HlsReader = class extends Readable {
 
   }
 
-  async process(segmentInfo, done) {
+  /**
+   * @param {import("hls-segment-reader").HlsStreamerObject} segmentInfo
+   */
+  async process(segmentInfo) {
 
-    let result;
     try {
       this.isReading = true;
 
       let stream;
       try {
-        stream = await this.decrypt(segmentInfo.stream, segmentInfo.segment && segmentInfo.segment.details.keys);
+        stream = await this.decrypt(segmentInfo.stream, segmentInfo.segment && segmentInfo.segment.entry.keys);
       }
       catch (err) {
         console.error('decrypt failed', err.stack);
@@ -115,6 +127,8 @@ const HlsReader = class extends Readable {
       const dispatcher = new Pati.EventDispatcher(stream);
       dispatcher.on('end', Pati.EventDispatcher.end);
 
+      const startTime = +segmentInfo.segment.entry.program_time + this.presentDelay * 1000;
+
       if (!this.isHooked) {
         // pull data and detect if we need to hook before end
         let buffered = 0;
@@ -122,7 +136,7 @@ const HlsReader = class extends Readable {
 
           buffered += chunk.length;
           if (!this.isHooked && buffered >= this.bufferSize) {
-            this.hook();
+            this.hook(startTime);
           }
         });
       }
@@ -137,18 +151,15 @@ const HlsReader = class extends Readable {
       }
 
       this.isReading = false;
-      this.hook();
+      this.hook(startTime);
     }
     catch (err) {
       console.error('process error', err.stack || err);
-      result = err;
-    }
-    finally {
-      done(result);
+      throw err;
     }
   }
 
-  hook() {                          // the hook is used to prebuffer
+  hook(startTime) {                          // the hook is used to prebuffer
 
     if (this.isHooked) return;
 
@@ -156,7 +167,7 @@ const HlsReader = class extends Readable {
 
     let s = this.buffer;
     if (this.sync) {
-      let smooth = TsSmooth();
+      let smooth = TsSmooth({ startTime });
       smooth.on('unpipe', () => {
 
         this.unpipe();
@@ -185,9 +196,8 @@ const HlsReader = class extends Readable {
   }
 };
 
-const hlsreader = module.exports = function hlsreader(segmentReader, options) {
+
+export default function hlsreader(segmentReader, options) {
 
   return new HlsReader(segmentReader, options);
 };
-
-hlsreader.HlsReader = HlsReader;
