@@ -2,7 +2,7 @@
 
 const debug = require('debug')('hls:tssmooth');
 
-const { Transform } = require('readable-stream');
+const { Transform } = require('stream');
 
 
 const internals = {};
@@ -97,6 +97,7 @@ const TsSmooth = class extends Transform {
     super({ highWaterMark: packetSize });
 
     this.packetSize = packetSize; // size of output packets
+    this.startTime = options.startTime;
 
     this.buffer = Buffer.alloc(0);
 
@@ -104,7 +105,7 @@ const TsSmooth = class extends Transform {
     this.last = null;
 
     this.errorLimit = 80000; /* 80 ms */
-    this.bitrate = 10E06;
+    //this.bitrate = 10E06;
     this.pcrTime = -1;
 
     this.pcrDelta = (pcr, lastPcr) => {
@@ -119,6 +120,11 @@ const TsSmooth = class extends Transform {
       if (this.pcr === -1) {
         this.pcr = pcr;
         this.last = internals.utime();
+        if (this.startTime) {
+          const delay = this.startTime - Date.now();
+          console.log('DELAY', this.startTime - Date.now())
+          this.last += delay * 1000;
+        }
       }
 
       let pcrDelta = this.pcrDelta(pcr, this.pcr);
@@ -198,7 +204,7 @@ const TsSmooth = class extends Transform {
           let error = (out.time - now) - out.delta;
           let waittime = (error > this.errorLimit) ? (error / 1000 - 5) : 0;
 
-          if (error < -2 * 1E6 || error > 300 * 1E6) {
+          if (!this.startTime && error < -2 * 1E6 || error > 300 * 1E6) {
             // negative == buffer too late
             // positive == buffer too early
             this.emit('warning', new RateError('PCR sync offset ' + (error / 1E6).toFixed(2) + 's error'));
@@ -206,6 +212,7 @@ const TsSmooth = class extends Transform {
             waittime = 0;
           } else if (error < -this.errorLimit) {
             // ignore the data since it is too late
+            startIndex = index;
             return setImmediate(processNext);
           }
 
